requirements.txt
gymnasium==0.29.1
numpy>=1.21.0
torch>=2.0.0
pyyaml>=6.0
matplotlib>=3.5.0
pillow>=8.0.0
tensorboard>=2.11.0